refactor(android-demo): consolidate movable-light into the Lighting demo (#1444)
The Samples tab carried two near-identical lighting cards — `lighting`
("Light Types") and `movable-light` ("Movable Light") — both staging the
same helmet model on the same topic. Merge them into a single Lighting
demo with an in-demo segmented-button mode switch: "Light Types"
(directional / point / spot, intensity, colour) and "Movable Light"
(drag to orbit the light). Every control from both demos is preserved —
this removes a demo entry, not a feature.

The retired `sceneview://demo/movable-light` deep link keeps working: a
new DeepLinkRouter.DEMO_ID_ALIASES table redirects retired ids to the
consolidated demo, and DeepLinkRouterTest pins the invariant that every
alias target is a live registered demo.

Conservative single-merge scope per #1444's open-ended ask — iOS keeps
its own MovableLightDemo (platform: android issue).
