+++ agents/sceneview/references/cheatsheet.md
@@ -77,8 +77,8 @@ There is **no `rememberARSession`** helper — configure via the
 | Node | Where verified | Notes |
 | --- | --- | --- |
 | `ModelNode(modelInstance, scaleToUnits, centerOrigin, isEditable)` | `ModelViewerDemo.kt` | Render a GLB. `isEditable = true` enables drag/scale/rotate |
-| `LightNode(type, intensity, direction, position, color, apply = { … })` | `MovableLightDemo.kt` | `type` is `LightManager.Type.POINT / SPOT / DIRECTIONAL / FOCUSED_SPOT / SUN` |
-| `CubeNode / SphereNode / CylinderNode / PlaneNode / CapsuleNode / TorusNode / ConeNode` | `GeometryDemo.kt`, `MovableLightDemo.kt` (Sphere) | Each takes `materialInstance` and shape-specific dimensions (e.g. `SphereNode(radius, ...)`) |
+| `LightNode(type, intensity, direction, position, color, apply = { … })` | `LightingDemo.kt` | `type` is `LightManager.Type.POINT / SPOT / DIRECTIONAL / FOCUSED_SPOT / SUN` |
+| `CubeNode / SphereNode / CylinderNode / PlaneNode / CapsuleNode / TorusNode / ConeNode` | `GeometryDemo.kt`, `LightingDemo.kt` (Sphere) | Each takes `materialInstance` and shape-specific dimensions (e.g. `SphereNode(radius, ...)`) |
 | `BillboardNode` | `BillboardDemo.kt` | Always faces the camera |
 | `ImageNode` | `ImageDemo.kt` | 2D image quad |
 | `TextNode` | `TextDemo.kt` | 3D text (uses `widthMeters` / `heightMeters`, NOT `scaleToUnits`) |
