+++ changelog.d/1444-consolidate-demos.md
@@ -1,0 +1,2 @@
+<!-- category: Changed -->
+- **Consolidated the two lighting demos into one ([#1444](https://github.com/sceneview/sceneview/issues/1444)).** The Android demo's `lighting` ("Light Types") and `movable-light` ("Movable Light") cards were near-identical — same helmet model, same topic — so they are merged into a single **Lighting** demo with an in-demo segmented-button mode switch: *Light Types* (directional / point / spot, intensity, colour) and *Movable Light* (drag to orbit the light). No feature is lost — every control from both demos is still present. The Samples tab now carries one lighting entry instead of two. The retired `sceneview://demo/movable-light` deep link keeps working: `DeepLinkRouter` aliases it to `lighting` via a new `DEMO_ID_ALIASES` table.
