--- samples/android-demo/src/main/res/values/strings_demo_movable_light.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `movable-light` demo (registered by
-     `MovableLightFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_movable_light_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_movable_light_title">Movable Light</string>
-    <string name="demo_movable_light_subtitle">Drag to orbit the light around the model</string>
-</resources>
